Extract mobile sort picker from PLP
